+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,7 +1,6 @@
 package org.apache.catalina;
 
 import jakarta.servlet.http.HttpSession;
-
 import java.io.IOException;
 
 /**
