+++ tomcat/src/main/java/org/apache/catalina/startup/Tomcat.java
@@ -1,10 +1,9 @@
 package org.apache.catalina.startup;
 
+import java.io.IOException;
 import org.apache.catalina.connector.Connector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
 
 public class Tomcat {
