+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,14 +4,10 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.nio.file.Watchable;
-import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.HttpRequestHandler;
+import org.apache.coyote.http11.request.model.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
