+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestHandler.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
+
+import org.apache.coyote.http11.request.model.HttpRequest;
 
 public class HttpRequestHandler {
