+++ tomcat/src/main/java/org/apache/coyote/http11/request/model/HttpVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request.model;
 
 import java.util.Arrays;
