+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpRequestHandlerTest.java
@@ -1,8 +1,11 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import org.apache.coyote.http11.request.model.HttpRequest;
+import org.apache.coyote.http11.request.model.HttpMethod;
+import org.apache.coyote.http11.request.model.HttpVersion;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
