+++ tomcat/src/test/java/org/apache/coyote/http11/request/model/HttpMethodTest.java
@@ -1,7 +1,8 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.request.model.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
